- Suppressed compilation warnings (unsigned and signed mismatches)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@6793 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/mob.c
@@ -1761,7 +1761,7 @@ int mob_dead(struct mob_data *md, struct block_list *src, int type)
 			continue;
 		temp++;
 
-		if(mvp_damage<md->dmglog[temp].dmg){
+		if(mvp_damage<(unsigned int)md->dmglog[temp].dmg){
 			third_sd = second_sd;
 			second_sd = mvp_sd;
 			mvp_sd=tmpsd[temp];
